updated user pref methods now under the model
