+++ src/scoops/lib/core/services/user_account_service.dart
@@ -1,6 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:scoops/core/constants/database_collections.dart';
+import 'package:scoops/core/data/user/user_preferences.dart';
 
 class UserAccountService {
   Future<UserPreferences> getUserSettings() async {
@@ -8,7 +9,7 @@ class UserAccountService {
         .collection(DatabaseCollections.userProfiles)
         .doc(FirebaseAuth.instance.currentUser.uid)
         .get();
-    return doc.exists ? mapToPreferences(doc) : null;
+    return doc.exists ? UserPreferences.mapFromDoc(doc) : null;
   }
 
   Future createEmptyPrefs() async {
@@ -19,48 +20,11 @@ class UserAccountService {
         .set(_emptyPrefs.convertToDocMap());
   }
 
-  UserPreferences mapToPreferences(DocumentSnapshot doc) {
-    var amenities = doc['amenities'] as List;
-    var groups = doc['beverage_groups'] as List;
-    var establishmentTypes = doc['establishment_types'] as List;
-    return UserPreferences(
-        amenities: amenities.map((e) => e as String).toList(),
-        beverageGroups: groups.map((e) => e as String).toList(),
-        establishmentTypes:
-            establishmentTypes.map((e) => e as String).toList());
-  }
-
   Future<bool> savePreferences(UserPreferences prefs) async {
-    try {} catch (e) {}
     await FirebaseFirestore.instance
         .collection(DatabaseCollections.userProfiles)
         .doc(FirebaseAuth.instance.currentUser.uid)
         .update(prefs.convertToDocMap());
     return true;
-  }
-
-  Map<String, List<dynamic>> mapPrefs(UserPreferences) => {};
-}
-
-class UserPreferences {
-  List<String> amenities;
-  List<String> establishmentTypes;
-  List<String> beverageGroups;
-
-  UserPreferences(
-      {this.amenities, this.beverageGroups, this.establishmentTypes});
-
-  UserPreferences.empty() {
-    this.amenities = List<String>.empty();
-    this.establishmentTypes = List<String>.empty();
-    this.beverageGroups = List<String>.empty();
-  }
-
-  Map<String, List<dynamic>> convertToDocMap() {
-    return {
-      'amenities': this.amenities,
-      'establishment_types': this.establishmentTypes,
-      'beverage_groups': this.beverageGroups
-    };
   }
 }
